+++ src/vscode-badge/index.ts
@@ -1,0 +1,1 @@
+export {VscodeBadge} from './vscode-badge';
